pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "portfolio-manager"
version = "0.1.0"
description = "Multi-repo agentic workflow orchestrator"
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary>=2.9.9",
    "PyGithub>=2.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
]

[project.scripts]
portfolio-manager = "portfolio_manager.__main__:main"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
